+++ docs/whats-new.md
@@ -8,6 +8,10 @@ Target Release Date: TBD (alpha releases will be made available)
 
 - `parseInBatches` can now be called on all loaders. Non-batched loaders will just return a single batch.
 
+**@loaders.gl/polyfills**
+
+- Improved robustness and error handling in Node.js when calling the `fetch` polyfill on unreadable or non-existent files. Underlying errors (`ENOEXIST`, `EISDIR` etc) are now caught and reported in `Response.statusText`.
+
 ## v2.2
 
 Release Date: June 18, 2020